Moved handgrading folder into project_view folder.
