+++ src/components/project_view/handgrading/file_panel.vue
