+++ src/components/project_view/handgrading/group_summary_panel.vue
