+++ tests/test_project_view/test_handgrading/test_handgrading_container.ts
@@ -5,10 +5,10 @@ import { Wrapper, WrapperArray } from '@vue/test-utils';
 import * as ag_cli from 'ag-client-typescript';
 import * as sinon from 'sinon';
 
-import GroupSummaryPanel from '@/components/handgrading/group_summary_panel.vue';
-import Handgrading from '@/components/handgrading/handgrading.vue';
-import HandgradingContainer from '@/components/handgrading/handgrading_container.vue';
-import { HandgradingStatus } from '@/components/handgrading/handgrading_status';
+import GroupSummaryPanel from '@/components/project_view/handgrading/group_summary_panel.vue';
+import Handgrading from '@/components/project_view/handgrading/handgrading.vue';
+import HandgradingContainer from '@/components/project_view/handgrading/handgrading_container.vue';
+import { HandgradingStatus } from '@/components/project_view/handgrading/handgrading_status';
 
 import * as data_ut from '@/tests/data_utils';
 import { managed_mount } from '@/tests/setup';
